Limit simEQ vars to 5, increase input width, clear output on var change
This five variable limit for the simultaneous equation calculator is due 
to overcrowding in the buttons when more variables are introduced.

Input width was increased to make the input text more readable, as it 
was too small before.

Output is now cleared on matrix size increase created a cleaner 
experience overall.